code clean up, some assets and shi
- Introduce MENU2 music type for randomized menu music selection
- Remove unused spaz character audio files and explosion variants
- Refactor Snowgrave class to properly inherit from bs.Actor with improved node handling and messaging
- Add SnowgraveTouchedMessage dataclass for cleaner collision handling
- Consolidate window renaming logic from bombsquda to delta startup module
- Fix Snowgrave instantiation in bomb.py with proper autoretain() call
- Enhance ParticalFactory with dedicated snowgrave_mat property
- Comment out unused explosion sound variants in bomb factory

+++ ba_data/python/delta/actor/snowgrave.py
@@ -1,107 +1,99 @@
-
-import bascenev1 as bs, random
-from delta.actor.particals import Partical, ParticalFactory
+"""Class for a snowgrave actor."""
+from typing import override, Any
+import bascenev1 as bs
+import babase as ba
+import random
+from delta.actor.particals import Partical, ParticalFactory, SnowgraveTouchedMessage
 from delta.actor.damagetext import DamageText
 from bascenev1lib.gameutils import SharedObjects
-
-
 
 class SnowgraveProp(bs.Actor):
     def __init__(self, position, meshes):
         super().__init__()
 
-class Snowgrave:
-    """
-        Cast a snowgrave in this position. 
+class Snowgrave(bs.Actor):
+    """Cast a snowgrave in this position. 
 
-        This creates an area that makes bomb icey
-        Slowly damages Spaz's and freezes them and eventually snowgraves them
-        (see class above)
-
+    This creates an area that makes bomb icey
+    Slowly damages Spaz's and freezes them and eventually snowgraves them
+    (see class above)
     """
-    def __init__(self, position, source_player):
-        
+    def __init__(self, position: tuple, source_player: bs.Player):
+        super().__init__()
         self.position = position
         self.source_player = source_player
         self.active = True
-        self.snowgrave_mat = bs.Material()
-        self.snowgrave_mat.add_actions(
-            conditions=
-                ('they_have_material', SharedObjects.get().object_material)
-            ,
-            actions=(
-                ('modify_part_collision', 'collide', True),
-                ('modify_part_collision', 'physical', False),
-                (
-                    'call',
-                    'at_connect',
-                    lambda: self.handle_touched_snowgrave(True)
-                ),
-                (
-                    'call',
-                    'at_disconnect',
-                    lambda: self.handle_touched_snowgrave(False)
-                ),
-            ),
-        )
+        self.loc = None
+        self.snowgrave_mat = ParticalFactory.get().snowgrave_mat
         ParticalFactory.get().snowgrave_sfx.play(2.5)
-        self.light = bs.newnode('light', attrs={
+        self.light = bs.newnode('light', delegate=self, attrs={
             'position': position,
             'color': (0.2, 0.5, 1.0),
             'radius': 0.2,
         })
+        # animate it in nicely
+        bs.animate(
+            self.light,
+            'radius',
+            {
+                0: 0,
+                0.2: 0.2,
+            }
+        )
         self.nodes: list[bs.Node] = []
         self.spaz_specific_hits = {}
         self.spaz_touch_cooldown = bs.time()
         bs.timer(1, self.make)
-        self.tick_timer = bs.timer(0.1, self.handle_nodes_inside, repeat=True)
+        self.tick_timer = bs.Timer(0.1, self.handle_nodes_inside, repeat=True)
 
     def handle_nodes_inside(self):
         if not self.active:
             return
         from bascenev1lib.actor.spaz import Spaz
-        for node in self.nodes:
-            if node.getdelegate(Spaz):
+        # wanna clone here cuz we're iterating it somewhat
+        for node in self.nodes[:]:
+            # node doesn't exist, die
+            if not node:
+                self.nodes.remove(node)
+                return
+            delegate = node.getdelegate(bs.Actor)
+            if isinstance(delegate, Spaz):
                 # cooldown
                 if not (self.spaz_touch_cooldown - bs.time() < 0):
                     return
                 self.spaz_touch_cooldown = bs.time() + 0.1
+                
+                actor_id = id(delegate)
 
-                actor = node.getdelegate(Spaz)
-                assert isinstance(actor, Spaz)
-                actor_id = id(actor)
-
-                actor.handlemessage(bs.HitMessage(
+                # keep hitting it
+                delegate.handlemessage(bs.HitMessage(
                     flat_damage=20,
                     source_player=self.source_player
                 ))
-                actor.impulse(y=15)
+                delegate.impulse(y=15)
                 if actor_id not in self.spaz_specific_hits:
                     self.spaz_specific_hits[actor_id] = 0
 
                 self.spaz_specific_hits[actor_id] += 1
-                print(self.spaz_specific_hits[actor_id])
 
-                if self.spaz_specific_hits[actor_id] > 15 and not actor.snowgraved:
-                    actor.snowgraved = True
-                   
+                # if this spaz got hit 15 times, freeze it
+                if self.spaz_specific_hits[actor_id] > 15 and not delegate.snowgraved:
+                    delegate.snowgraved = True
+                    delegate.handlemessage(bs.FreezeMessage())
 
-            else:
-                if node.getdelegate(bs.Actor):
-                    actor = node.getdelegate(bs.Actor)
-                    assert isinstance(actor, bs.Actor)
-                    actor.impulse(y=40)
-
-                    
-
+            elif isinstance(delegate, bs.Actor):
+                # any other actor gets swept away... hopefully
+                delegate.impulse(y=40)
 
 
     def handle_touched_snowgrave(self, entered):
-
-        
-
         collision = bs.getcollision()
-        node = collision.opposingnode
+        # try getting the node, but if it
+        # doesnt exist then just nullify it
+        try:
+            node = collision.opposingnode
+        except ba._error.NodeNotFoundError:
+            node = None
 
         if not node:
             return
@@ -110,27 +102,28 @@ class Snowgrave:
         else:
             self.nodes.remove(node)
 
-        
-        
     
     def make(self):
-
         hitbox_size = (2, 10, 2)
-        self.loc=bs.newnode('locator',
-                    attrs={'shape': 'box',
-                           'position': self.position,
-                           'color': (1,1,1),
-                           'opacity': 1.0,
-                           'draw_beauty': True,
-                           'size': hitbox_size,
-                           'additive': False})
+        # i'm assuming this is debug;
+        # if not, just keep it true
+        debug = False
+        if debug:
+            self.loc=bs.newnode('locator',
+                        attrs={'shape': 'box',
+                               'position': self.position,
+                               'color': (1,1,1),
+                               'opacity': 1.0,
+                               'draw_beauty': True,
+                               'size': hitbox_size,
+                               'additive': False})
                            
         
-        self.hitbox=bs.newnode('region',
+        self.hitbox=bs.newnode('region',delegate=self,
                     attrs={'scale': hitbox_size,
                            'position': self.position,
                            'type': 'box',
-                           'materials': [self.snowgrave_mat,SharedObjects.get().attack_material]})
+                           'materials': [self.snowgrave_mat, SharedObjects.get().attack_material]})
         timer=0.1
         for _ in range(30):
             bs.timer(timer, self.emit)
@@ -140,15 +133,18 @@ class Snowgrave:
     
     def end(self):
         self.active = False
-        self.light.delete()
-        self.hitbox.delete()
-        self.loc.delete()
-        self.tick_timer = None
+        # delete nodes if they exist
+        if self.light:
+            self.light.delete()
+        if self.hitbox:
+            self.hitbox.delete()
+        if self.loc:
+            self.loc.delete()
+        if self.tick_timer:
+            self.tick_timer = None
         self.spaz_specific_hits = {}
         self.nodes.clear()
-        
 
-        
     def emit(self):
         bs.emitfx(
             position=self.position,
@@ -159,6 +155,7 @@ class Snowgrave:
                 20, 
                 0
             ),
+            scale=1.3,
             chunk_type='ice',
         )
         bs.emitfx(
@@ -171,6 +168,7 @@ class Snowgrave:
                 random.uniform(-0.2, 0.2)
             ),
             chunk_type='ice',
+            scale=1.2,
         )
         if random.randint(0, 1) == 0:
             p=Partical(
@@ -203,4 +201,21 @@ class Snowgrave:
             0.1: 0.65,
             0.2: 0.2
         })
-        
+    
+    @override
+    def handlemessage(self, msg):
+        if isinstance(msg, bs.DieMessage):
+            self.end()
+        elif isinstance(msg, SnowgraveTouchedMessage):
+            # just fallback to our function
+            self.handle_touched_snowgrave(msg.state)
+        else:
+            return super().handlemessage(msg)
+        return None
+    
+    @override
+    def exists(self) -> bool:
+        # should use something else 
+        # than the light maybe??
+        # REMINDER: this gets used for autoretain
+        return bool(self.light)
